Add files via upload
The system should be designed in such a way that only authorized people should be allowed to access some particular modules. The records should be modified by only authenticated person and no one else. The user should always be in control of the application and not the vice versa. 
The user interface should be consistent so that the user can handle the application with ease and speed. The application should be visually, conceptually clear.